Split package manager impl
